+++ src/database/queries/books_table.sql
@@ -1,17 +1,23 @@
 CREATE TABLE IF NOT EXISTS books (
-	id SERIAL PRIMARY KEY,
-	name varchar(100) NOT NULL,
-	author varchar(50) NOT NULL REFERENCES authors (name),
-	due_date date DEFAULT NULL,
-	pages smallint UNSIGNED DEFAULT NULL,
-	genre varchar(20) DEFAULT NULL,
-	rating char(1) DEFAULT NULL,
-	viewers_counter int UNSIGNED DEFAULT 0,
-	downloads_counter int UNSIGNED DEFAULT 0,
-	epub bytea DEFAULT NULL,
-	epub_size bigint UNSIGNED DEFAULT NULL,
-	pdf bytea DEFAULT NULL,
-	pdf_size bigint UNSIGNED DEFAULT NULL,
+	id                SERIAL,
+	name              varchar(100) NOT NULL,
+	author            varchar(50)  NOT NULL
+	REFERENCES authors (name) ON DELETE SET NULL (author),
+	duedate           date         DEFAULT NULL,
+	language          char(2)      DEFAULT NULL,
+	pages             smallint     DEFAULT NULL,
+	genre             varchar(20)  DEFAULT NULL,
+	description       text         DEFAULT NULL,
+	wikilink          text         DEFAULT NULL,
+	rating            char(1)      DEFAULT NULL,
+	viewerscounter    int          DEFAULT 0,
+	downloadscounter  int          DEFAULT 0,
+	epub              bytea        DEFAULT NULL,
+	epubsize          bigint       DEFAULT 0,
+	pdf               bytea        DEFAULT NULL,
+	pdfsize           bigint       DEFAULT 0,
+	-- ------------------------constraints-------------------------------------
+	PRIMARY KEY (id),
 	-- same author cannot write same book twice, right?
-	UNIQUE (name, author);
-)
+	UNIQUE (name, author)
+);