+++ src/database/queries/increment_book_view.sql
@@ -1,0 +1,5 @@
+-- increment book page views counter
+
+UPDATE authors
+SET viewerscounter = viewerscounter + 1
+WHERE id = %s AND name = %s;