+++ src/database/queries/select_some_books.sql
@@ -1,0 +1,6 @@
+-- select a list of books
+
+SELECT name, author, language
+FROM books
+ORDER BY rating DESC
+LIMIT 15;